+++ src/components/start-airdrop/UploadNFTs/components/AddNftButton.tsx
@@ -1,4 +1,5 @@
 import { Box, Button, Text } from "@chakra-ui/react"
+import { motion } from "framer-motion"
 import { FilePlus } from "phosphor-react"
 import { ReactElement } from "react"
 import { DropzoneInputProps, DropzoneRootProps } from "react-dropzone"
@@ -14,22 +15,24 @@ const AddNftButton = ({
   inputProps,
   isDragActive,
 }: Props): ReactElement => (
-  <Box width="full" minHeight="sm" height="full" position="relative">
-    <input id="dropzone" {...inputProps} hidden />
-    <Button
-      {...dropzoneProps}
-      as="label"
-      htmlFor="dropzone"
-      width="full"
-      height="full"
-      colorScheme="yellow"
-      variant="outline"
-      leftIcon={<FilePlus size={30} />}
-      aria-label="Upload images for NFTs"
-    >
-      {isDragActive ? <Text>Drop the files here</Text> : <Text>Add image(s)</Text>}
-    </Button>
-  </Box>
+  <motion.div whileTap={{ scale: 0.95 }} style={{ cursor: "pointer" }}>
+    <Box width="full" minHeight="sm" height="full" position="relative">
+      <input id="dropzone" {...inputProps} hidden />
+      <Button
+        {...dropzoneProps}
+        as="label"
+        htmlFor="dropzone"
+        width="full"
+        height="full"
+        colorScheme="yellow"
+        variant="outline"
+        leftIcon={<FilePlus size={30} />}
+        aria-label="Upload images for NFTs"
+      >
+        {isDragActive ? <Text>Drop the files here</Text> : <Text>Add image(s)</Text>}
+      </Button>
+    </Box>
+  </motion.div>
 )
 
 export default AddNftButton
