+++ src/components/start-airdrop/UploadNFTs/components/NftCard/components/TraitInput.tsx
