+++ src/components/start-airdrop/UploadNFTs/components/NftCard/index.tsx
@@ -1,0 +1,3 @@
+import NftCard from "./NftCard"
+
+export default NftCard
