--- src/components/start-airdrop/UploadNFTs/components/RoleCard/index.tsx
@@ -1,3 +1,0 @@
-import RoleCard from "./RoleCard"
-
-export default RoleCard
